Enable exploration by default
Removed the old cart as well

In support of: #3211
