+++ frontend/src/pages/cart/components/existing-exploration-form.tsx
@@ -6,9 +6,9 @@ import { useSelector } from 'react-redux';
 import { useNavigate } from 'react-router';
 import { PosthogEvents } from '@/constants/posthog.constants.ts';
 import { useFormPersist } from '@/hooks/use-form-persist.tsx';
-import { FormIssues } from '@/pages/cart-explorations/components/form-item-issues.tsx';
-import { JustificationFormItem } from '@/pages/cart-explorations/components/form-item-justification.tsx';
-import { useCartOverlapCheck } from '@/pages/cart-explorations/hooks/use-cart-overlap-check.ts';
+import { FormIssues } from '@/pages/cart/components/form-item-issues.tsx';
+import { JustificationFormItem } from '@/pages/cart/components/form-item-justification.tsx';
+import { useCartOverlapCheck } from '@/pages/cart/hooks/use-cart-overlap-check.ts';
 import { ExplorationTabKeys } from '@/pages/exploration/exploration-tab-keys.ts';
 import { useAppDispatch } from '@/store';
 import { selectCurrentUser } from '@/store/api/services/auth-slice.ts';
