+++ frontend/src/pages/cart/components/new-data-product-form.tsx
@@ -9,7 +9,7 @@ import styles from '@/components/data-products/data-product-form/data-product-fo
 import { DataProductFormItems } from '@/components/data-products/data-product-form/data-product-form-items.component.tsx';
 import { PosthogEvents } from '@/constants/posthog.constants.ts';
 import { useFormPersist } from '@/hooks/use-form-persist.tsx';
-import { JustificationFormItem } from '@/pages/cart-explorations/components/form-item-justification.tsx';
+import { JustificationFormItem } from '@/pages/cart/components/form-item-justification.tsx';
 import { TabKeys as DataProductTabKeys } from '@/pages/data-product/components/data-product-tabs/data-product-tabkeys.ts';
 import { useAppDispatch } from '@/store';
 import { selectCurrentUser } from '@/store/api/services/auth-slice.ts';
