+++ frontend/src/pages/cart/hooks/use-cart-overlap-check.ts
