+++ frontend/src/routes.tsx
@@ -3,10 +3,8 @@ import { createBrowserRouter, Navigate, RouterProvider, useParams } from 'react-
 import { AuthLayout } from '@/components/layout/auth/auth.layout.tsx';
 import PublicLayout from '@/components/layout/public/public.layout.tsx';
 import RootLayout from '@/components/layout/root/root.layout.tsx';
-import { AppConfig } from '@/config/app-config.ts';
 import { Logout } from '@/pages/auth/logout/logout-page.tsx';
 import Cart from '@/pages/cart/cart.page.tsx';
-import ExplorationsCart from '@/pages/cart-explorations/cart.page.tsx';
 import { DataProduct } from '@/pages/data-product/data-product.page.tsx';
 import { DataProductCreate } from '@/pages/data-product-create/data-product-create.page.tsx';
 import { DataProductEdit } from '@/pages/data-product-edit/data-product-edit.page.tsx';
@@ -103,7 +101,7 @@ const router = createBrowserRouter([
                         path: ApplicationPaths.MarketplaceCart,
                         children: [
                             {
-                                element: AppConfig.ExplorationsEnabled() ? <ExplorationsCart /> : <Cart />,
+                                element: <Cart />,
                                 index: true,
                             },
                         ],
